+++ data_parser/measuredata.h
@@ -6,14 +6,14 @@
 class MeasureData
 {
 public:
-    MeasureData(const long double _timeStamp, const long double _value);
+    MeasureData(const double _timeStamp, const double _value);
 
     void printAll() const;
     QPointF getPoint() const;
 
 private:
-    long double m_timeStamp;
-    long double m_value;
+    double m_timeStamp;
+    double m_value;
 };
 
 #endif // MEASUREDATA_H
